Add 'smart' input ordering
We want to balance a number of things when selecting inputs.

* Consuming small inputs to ensure large amounts of tiny inputs don't build up, requiring fusions and large transaction sizes
* Ensuring not only small inputs are selected, so the transaction is not too large to be sent
* Consistently selecting roughly the same amount of inputs for a transaction of the same amount, so fee per byte estimates are reasonably accurate

To satisfy this, we select inputs in the following manner:
We sort our inputs into base10 buckets. That is, 1, 2, 3, 4, .. 9 will all be in a bucket, 10..90 will be in another, 100..900 another, and so on.
Next, we loop through each bucket, taking the smallest input from each, then we begin the loop again.

Therefore, if we have inputs like the following: [1, 1, 1, 2, 5, 10, 20, 30, 100, 200, 300, 1000, 1000, 2000]
First, they will be sorted like so: [[1, 1, 1, 2, 5], [10, 20, 30], [100, 200, 300], [1000, 1000, 2000]]
Then, we will select inputs as follows: [1, 10, 100, 1000, 1, 20, 200, 1000, 1, 30, 300, 2000, 2, 5]
Finally, when we go to make the transaction, we select as many inputs as needed to cover the transaction amount.
For example, if we required 2500 units, we would end up using [1, 10, 100, 1000, 1, 20, 200, 1000, 1, 30, 300] = 2663