FIX: Resolve issue whereby the `status` command in the daemon does not always return a result
This is due to binding to a non-interface IP (0.0.0.0) and the fact that we cannot make HTTP calls to 0.0.0.0 reliably and should specify an IP address to issue the RPC calls against. This commit simply replaces 0.0.0.0 with 127.0.0.1 (localhost) for purposes of routing the client RPC request. This also moves this magic out into the daemon.cpp file to make it so that we're only passing in the ip and port for the RPC interface as we do not need a full pointer to the RpcServer actually in commands handler.